labelox/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "labelox"
version = "0.1.0"
description = "AI-powered annotation platform for computer vision datasets"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "sqlalchemy>=2.0.30",
    "pydantic>=2.7.0",
    "Pillow>=10.3.0",
    "numpy>=1.26.0",
    "opencv-python>=4.10.0",
]

[project.optional-dependencies]
desktop = [
    "PyQt6>=6.7.0",
]
ai = [
    "torch>=2.1.0",
    "ultralytics>=8.2.0",
    "transformers>=4.40.0",
    "scipy>=1.13.0",
]
web = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "websockets>=12.0",
]
similarity = [
    "faiss-cpu>=1.8.0",
]

[project.scripts]
labelox = "labelox.desktop.main:main"
labelox-web = "labelox.web.backend.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["labelox*"]
